# console/service/api/swagger.yaml
...
swagger: '2.0'
info:
  title: autobase console
  description: API for autobase console
  version: 2.11.0
host: localhost:8080
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
basePath: "/api/v1"
paths:
...
  /users:
    get:
      summary: "pg_genie: list console users (admin)"
      tags:
        - user
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
    post:
      summary: "pg_genie: add a local user (admin)"
      tags:
        - user
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/Request.UserCreate"
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/User"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
  /users/{id}:
    patch:
      summary: "pg_genie: change a user's role, display name or password (admin)"
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          type: integer
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/Request.UserUpdate"
      responses:
        '200':
          description: OK
          schema:
            $ref: "#/definitions/User"
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
    delete:
      summary: "pg_genie: remove a user (admin)"
      tags:
        - user
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '204':
          description: OK
        '400':
          description: Error
          schema:
            $ref: "#/definitions/Response.Error"
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      display_name:
        type: string
      role:
        type: string
        description: "Legacy: the user's group (admin, operator, viewer); same as attributes.group"
      attributes:
        type: object
        description: "ABAC attributes, e.g. {\"group\": \"operator\", \"team\": \"payments\"}"
        additionalProperties:
          type: string
      permissions:
        type: object
        description: "Only in /auth/me: {global: [permissions], clusters: {<id>: [permissions]}}"
      preferences:
        type: object
        description: "Only in /auth/me: the user's home page choices"
      auth_provider:
        type: string
      created_at:
        type: string
        format: date-time
      last_login_at:
        type: string
        format: date-time
        x-nullable: true
  Request.UserCreate:
    type: object
    required: [username, password]
    properties:
      username:
        type: string
      display_name:
        type: string
      password:
        type: string
      role:
        type: string
        description: "Legacy: sets the group attribute"
      attributes:
        type: object
        additionalProperties:
          type: string
  Request.UserUpdate:
    type: object
    properties:
      display_name:
        type: string
      password:
        type: string
      role:
        type: string
        description: "Legacy: sets the group attribute"
      attributes:
        type: object
        description: Replaces all attributes when given
        additionalProperties:
          type: string
      home:
        type: object
        description: "pg_genie: the user's home page, set by an admin (cards in order, start page)"
        properties:
          cards:
            type: array
            items:
              type: string
          start_page:
            type: string
...
  Response.Error:
    title: Error object
    type: object
    properties:
      code:
        type: integer
      title:
        type: string
      description:
        type: string
...